refactor: time 패키지 reservationTime으로 변경
