+++ src/main/java/roomescape/reservation/dto/ReservationResponse.java
@@ -2,7 +2,7 @@ package roomescape.reservation.dto;
 
 import java.time.LocalDate;
 import roomescape.reservation.domain.Reservation;
-import roomescape.time.dto.ReservationTimeResponse;
+import roomescape.reservationTime.dto.ReservationTimeResponse;
 
 public record ReservationResponse(Long id, String name, LocalDate date, ReservationTimeResponse time) {
     public static ReservationResponse from(final Reservation reservation) {
